requirements.txt
boto3
python-dotenv
streamlit
plotly
pandas
